+++ doc/changelog.md
@@ -1,2 +1,3 @@
 ### v0.0.1
 - New: barebones project template
+- New: add Point and Rectangle